Cleaning import routes
